Update RUC#s and Organize Tabs
This script that sets the value of cell C4 on all sheets in the active spreadsheet to the respective sheet name, except for the sheet named "Dashboard". Further, This script sorts the sheets in a Google Sheets spreadsheet based on the last 4 digits of the tab name in a specific format ("XXX-0001 (XXX)"). It first sorts the sheets based on the tab name and then reorders the sheets based on the sorted array.